feat(242): move language keys
