+++ ui/pages/profile/ChangeLanguagePage/ChangeLanguagePage.tsx
@@ -1,5 +1,5 @@
 import { Routes } from '@/modules/navigation/domain/entities/routes';
-import { Languages } from '@/modules/translations/domain/entities/Languages';
+import { Languages } from '@/modules/shared/domain/Languages';
 import { BasicView } from '@/ui/components/view/BasicView/BasicView';
 import { View } from 'react-native';
 import { useChangeLanguagePageLogic } from './ChangeLanguagePage.logic';
